[WIP] Fix employee termination bugs
Switches all dates to uint64 (for consistency with the rest of
aragon-apps/aragonOS).
Adds employee `endDate` and `accruedValue`.
Closes #389 and #390.

TODO:
- [ ] Add tests for new logic and bring coverage back up to 100%

+++ future-apps/payroll/test/mocks/PayrollMock.sol
@@ -4,10 +4,10 @@ import "../../contracts/Payroll.sol";
 
 
 contract PayrollMock is Payroll {
-    uint private _mockTime = now;
+    uint64 private _mockTime = uint64(now);
 
-    function mockUpdateTimestamp() public { _mockTime = now; }
-    function mockSetTimestamp(uint i) public { _mockTime = i; }
-    function getTimestampPublic() public constant returns (uint256) { return _mockTime; }
-    function getTimestamp() internal constant returns (uint256) { return _mockTime; }
+    function mockUpdateTimestamp() public { _mockTime = uint64(now); }
+    function mockSetTimestamp(uint64 i) public { _mockTime = i; }
+    function getTimestampPublic() public constant returns (uint64) { return _mockTime; }
+    function getTimestamp() internal constant returns (uint64) { return _mockTime; }
 }
